Logic Update
This update will organize the code more efficiently for debugging.  Refactored everything into an OOP structure and provided more logging for workflow
